[TOOL-3120] Redirect to last visited team page if authenticated (#6028)
