+++ apps/dashboard/src/app/team/components/last-visited-page/consts.ts
@@ -1,0 +1,3 @@
+// Only keep constants here - this file is imported in middleware too
+
+export const LAST_VISITED_TEAM_PAGE_PATH = "last-visited-team-page-path";
